refactor(api): segment the local-lane surface out of the Operations port
The Operations port declared all 47 control-plane methods, but the
in-container local-socket lane (LocalApiRouter) uses only 15 of them —
the spec CLI's global-spec surface and board, spec-room messaging, and a
run's own conversation and credential lane. The other 32 are web-only
(project lifecycle, dispatch, snapshots, agent status, run control, events,
reviews) and the local lane can never reach them.

Extract those 15 into a LocalLaneOperations role interface and have
Operations extend it, then narrow LocalApiRouter and LocalApiSocket to
depend on the role rather than the full port. Interface segregation with a
real consumer: the in-container lane now compiles against exactly its
contract, so a web-only op is unreachable from it by construction, and the
local surface is documented in one place. No behavior change — the union of
the two interfaces is identical to the old port; the API lane suite and
mvn clean verify are green.
