+++ sail-infra/src/main/java/ai/singlr/sail/api/LocalApiSocket.java
@@ -60,7 +60,7 @@ public final class LocalApiSocket implements AutoCloseable {
   private volatile Thread acceptLoop;
   private volatile boolean closed;
 
-  public LocalApiSocket(EventBus bus, Operations operations, Path socketPath) {
+  public LocalApiSocket(EventBus bus, LocalLaneOperations operations, Path socketPath) {
     this(new LocalApiRouter(bus, operations), socketPath, DEFAULT_MAX_IN_FLIGHT);
   }
